move calculateTxSize to common/helpers/btc
